requirements/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-cors-middleware==1.3.1
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.3
drfdocs==0.0.11
idna==2.6
Markdown==2.6.9
oauthlib==2.0.6
py==1.5.2
pytest==3.2.5
pytz==2017.3
requests==2.18.4
urllib3==1.22
